requirements.txt
fastapi
uvicorn
requests
httpx
sqlalchemy
psycopg2-binary
python-dotenv
selenium
alembic 
pydantic 